# gravitee-apim-rest-api/gravitee-apim-rest-api-portal/gravitee-apim-rest-api-portal-rest/src/main/resources/portal-openapi.yaml
openapi: 3.0.0
info:
    title: Gravitee.io Portal Rest API
    description: API dedicated to the devportal part of Gravitee
    termsOfService: terms
    contact:
        name: The Gravitee Team
        url: https://gravitee.io
        email: contact@graviteesource.com
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0
    version: "4.11.0-SNAPSHOT"
servers:
    - url: /portal/environments/{envId}
      description: The portal API for a given environment
      variables:
          envId:
              default: DEFAULT
              description: ID of the environment
...
security:
    - BasicAuth: []
    - CookieAuth: []
paths:
...
    /subscriptions:
        get:
            tags:
                - Subscription
            parameters:
                - $ref: "#/components/parameters/apiIdQueryParam"
                - $ref: "#/components/parameters/applicationIdQueryParam"
                - $ref: "#/components/parameters/subscriptionStatusesQueryParam"
                - $ref: "#/components/parameters/pageNumberParam"
                - $ref: "#/components/parameters/pageSizeParam"
            description: |
                List all ACCEPTED, PAUSED & PENDING subscriptions, filtered by api and/or by application. At least an api or an application must be provided.

                User must have the APPLICATION_SUBSCRIPTION[READ] permission to list subscription with application query param.\
                User must have the API_SUBSCRIPTION[READ] permission to list subscription with api query param.

            operationId: getSubscriptions
            responses:
                200:
                    description: List of Subscriptions
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SubscriptionsResponse"
                400:
                    description: At least an api or an application must be provided.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                403:
                    $ref: "#/components/responses/PermissionError"
                500:
                    $ref: "#/components/responses/InternalServerError"
        post:
            tags:
                - Subscription
            requestBody:
                description: Use to create a subscription.
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/SubscriptionInput"
            summary: Create a subscription.
            description: |
                Create a new subscription. If the application used to subscribe has no api key mode (UNSPECIFIED) and the plan is APIKEY, then the api key mode can be set during creation.
                Otherwise, the api_key_mode field is ignored.

                User must have APPLICATION_SUBSCRIPTION[CREATE] permission.
            operationId: createSubscription
            responses:
                200:
                    description: Created subscription.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Subscription"
                400:
                    description: Input must not be null.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                403:
                    $ref: "#/components/responses/PermissionError"
                500:
                    $ref: "#/components/responses/InternalServerError"
    /subscriptions/{subscriptionId}:
        parameters:
            - $ref: "#/components/parameters/subscriptionIdParam"
        get:
            tags:
                - Subscription
            parameters:
                - $ref: "#/components/parameters/subscriptionIncludeQueryParam"

            summary: Get a subscription.
            description: |
                Get a subscription.

                User must have API_SUBSCRIPTION[CREATE] or APPLICATION_SUBSCRIPTION[CREATE] permission.
            operationId: getSubscriptionById
            responses:
                200:
                    description: One subscription.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Subscription"
                403:
                    $ref: "#/components/responses/PermissionError"
                404:
                    description: Subscription not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
        put:
            tags:
                - Subscription
            summary: Update a subscription
            requestBody:
                description: Configuration of the subscription to update
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateSubscriptionInput"
            description: |
                Update a subscription.

                User must have APPLICATION_SUBSCRIPTION[UPDATE] permission.
            operationId: updateSubscription
            responses:
                200:
                    description: Subscription updated
                403:
                    $ref: "#/components/responses/PermissionError"
                404:
                    description: Subscription not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                400:
                    description: Subscription not updatable, or bad subscription configuration format
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
components:
...
    parameters:
...
        subscriptionIdParam:
            name: subscriptionId
            in: path
            required: true
            description: Id of a subscription.
            schema:
                type: string
...
        apiIdQueryParam:
            name: apiId
            in: query
            required: false
            description: Id of an api.
            schema:
                type: string
        applicationIdQueryParam:
            name: applicationId
            in: query
            required: false
            description: Id of an application.
            schema:
                type: string
...
        pageNumberParam:
            name: page
            in: query
            required: false
            description: The page number for pagination.
            schema:
                type: integer
                default: 1
        pageSizeParam:
            name: size
            in: query
            required: false
            description: |
                The number of items per page for pagination.
                If the size is 0, the response contains only metadata and returns the values as for a non-paged resource.
                If the size is -1, the response contains all datas.
            schema:
                type: integer
                default: 10

        #Apis filtering
...
        subscriptionStatusesQueryParam:
            name: statuses
            in: query
            required: false
            description: status of subscription.
            schema:
                type: array
                items:
                    type: string
                    enum:
                        - ACCEPTED
                        - CLOSED
                        - PAUSED
                        - PENDING
                        - REJECTED
        subscriptionIncludeQueryParam:
            name: include
            in: query
            required: false
            description: Comma-separated list of related objects to include in the response.
            schema:
                type: array
                items:
                    type: string
                    enum:
                        - keys
                        - consumerConfiguration
...
    schemas:
...
        ErrorResponse:
            properties:
                errors:
                    description: List of errors
                    type: array
                    items:
                        $ref: "#/components/schemas/Error"
...
        SubscriptionsResponse:
            properties:
                data:
                    description: List of subscriptions.
                    type: array
                    items:
                        $ref: "#/components/schemas/Subscription"
                metadata:
                    $ref: "#/components/schemas/MetadataMap"
                links:
                    $ref: "#/components/schemas/Links"
...
        Error:
            properties:
                status:
                    description: Status HTTP of the error
                    type: string
                message:
                    description: Message of the error
                    type: string
                code:
                    description: Technical code of the error
                    type: string
                parameters:
                    description: Parameters
                    type: object
                    additionalProperties:
                        type: string
...
        Group:
            required:
                - id
                - name
            properties:
                id:
                    description: Unique identifier of a group.
                    type: string
                name:
                    description: Name of the group.
                    type: string
...
        Application:
            properties:
                id:
                    description: Unique identifier of an application.
                    type: string
                name:
                    description: Name of the application.
                    type: string
                description:
                    description: Description of the application.
                    type: string
                domain:
                    description: Domain used by the application.
                    type: string
                applicationType:
                    description: Type of the application (Web, Mobile, ...).
                    type: string
                origin:
                    description: Application origin. e.g., MANAGEMENT, KUBERNETES, INTEGRATION.
                    type: string
                hasClientId:
                    description: True if the application has client id.
                    type: boolean
                owner:
                    $ref: "#/components/schemas/User"
                created_at:
                    description: Creation date and time of the application.
                    type: string
                    format: date-time
                updated_at:
                    description: Last update date and time of the application.
                    type: string
                    format: date-time
                groups:
                    description: Array of groups associated to the application.
                    type: array
                    items:
                        $ref: "#/components/schemas/Group"
                picture:
                    description: |
                        Picture of the application. This attribute is only used to update a picture.\
                        To get the application picture, use /application/{applicationId}/picture.
                    type: string
                    writeOnly: true
                background:
                    description: |
                        Background of the application. This attribute is only used to update a picture.\
                        To get the application picture, use /application/{applicationId}/background.
                    type: string
                    writeOnly: true
                api_key_mode:
                    $ref: "#/components/schemas/ApiKeyModeEnum"
                settings:
                    $ref: "#/components/schemas/ApplicationSettings"
                _links:
                    $ref: "#/components/schemas/ApplicationLinks"
...
        Subscription:
            required:
                - id
                - api
                - application
                - plan
                - status
            properties:
                id:
                    description: Unique identifier of a subscription.
                    type: string
                api:
                    description: Subscribed API.
                    type: string
                application:
                    description: Subscribing application.
                    type: string
                plan:
                    description: Subscribed plan.
                    type: string
                request:
                    description: Comment of the user when asking for a subscription.
                    type: string
                reason:
                    description: Reason of the user when processing (accepting/rejecting) a subscription.
                    type: string
                created_at:
                    description: Creation date and time of the subscription.
                    type: string
                    format: date-time
                updated_at:
                    description: Updated date and time of the subscription.
                    type: string
                    format: date-time
                processed_at:
                    description: Date and time when the subscription request was processed.
                    type: string
                    format: date-time
                start_at:
                    description: Start date and time of the subscription.
                    type: string
                    format: date-time
                end_at:
                    description: Expiration date and time of the subscription.
                    type: string
                    format: date-time
                paused_at:
                    description: Paused date and time of the subscription.
                    type: string
                    format: date-time
                closed_at:
                    description: Closed date and time of the subscription.
                    type: string
                    format: date-time
                subscribed_by:
                    description: The user who subscribed.
                    type: string
                status:
                    description: Status of the subscription.
                    type: string
                    enum:
                        - PENDING
                        - ACCEPTED
                        - CLOSED
                        - REJECTED
                        - PAUSED
                consumerStatus:
                    description: Consumer status of the subscription.
                    type: string
                    enum:
                        - STARTED
                        - STOPPED
                        - FAILURE
                consumerPausedAt:
                    description: Paused date and time of the subscription for the customer.
                    type: string
                    format: date-time
                failureCause:
                    description: The cause of the failure
                    type: string
                origin:
                    description: The origin of the subscription
                    type: string
                    enum:
                        - KUBERNETES
                        - MANAGEMENT
                ###################
                # may be included #
                ###################
                keys:
                    description: |
                        Only returned with (*)/subscriptions/{subscriptionId}*.
                        Need *include* query param to contain 'keys'.

                        List of APIKeys of the subscription.
                    type: array
                    items:
                        $ref: "#/components/schemas/Key"
                consumerConfiguration:
                    $ref: "#/components/schemas/SubscriptionConsumerConfiguration"
...
        SubscriptionInput:
            properties:
                application:
                    description: Id of the application which subscribe
                    type: string
                plan:
                    description: Id of the plan to which the application subscribe
                    type: string
                request:
                    description: A request message to the api owner why a user want to subscribe
                    type: string
                general_conditions_accepted:
                    description: Specify if the general conditions have been accepted by the API subscriber
                    type: boolean
                general_conditions_content_revision:
                    description: Page revision of the general conditions visualized and accepted by the API subscriber
                    $ref: "#/components/schemas/PageRevisionId"
                metadata:
                    description: Subscription metadata
                    type: object
                    additionalProperties:
                        type: string
                configuration:
                    description: Subscription configuration
                    $ref: "#/components/schemas/SubscriptionConfigurationInput"
                api_key_mode:
                    $ref: "#/components/schemas/ApiKeyModeEnum"
        UpdateSubscriptionInput:
            properties:
                metadata:
                    description: Subscription metadata
                    type: object
                    additionalProperties:
                        type: string
                configuration:
                    description: Subscription configuration
                    $ref: "#/components/schemas/SubscriptionConfigurationInput"
        SubscriptionConfigurationInput:
            properties:
                channel:
                    description: The channel to subscribe to if any
                    type: string
                entrypointId:
                    description: The entrypointId to subscribe to if any
                    type: string
                entrypointConfiguration:
                    description: The entrypoint configuration if needed
                    type: object
...
        ApplicationSettings:
            properties:
                app:
                    $ref: "#/components/schemas/SimpleApplicationSettings"
                oauth:
                    $ref: "#/components/schemas/OAuthClientSettings"
                tls:
                    $ref: "#/components/schemas/TlsClientSettings"
        SimpleApplicationSettings:
            properties:
                type:
                    type: string
                client_id:
                    type: string
...
        TlsClientSettings:
            properties:
                client_certificate:
                    type: string
...
        Key:
            description: Subscription Key
            properties:
                id:
                    type: string
                key:
                    type: string
                application:
                    $ref: "#/components/schemas/Application"
                subscriptions:
                    type: array
                    items:
                        $ref: "#/components/schemas/Subscription"
                paused:
                    type: boolean
                revoked:
                    type: boolean
                expired:
                    type: boolean
                created_at:
                    type: string
                    format: date-time
                revoked_at:
                    type: string
                    format: date-time
                expire_at:
                    type: string
                    format: date-time
                hash:
                    type: string
                    description: The MD5 hash of the API Key
...
        ApiKeyModeEnum:
            description: |
                The API Key mode to use for this application.
                  - The `SHARED` API Key mode allows consumer to use the same API Key across all the subscriptions
                  - The `EXCLUSIVE` API Key mode will result to a new API Key being generated for each subscription
                  - The `UNSPECIFIED` API Key mode is a marker value informing that no choice as been made yet regarding
                    the API Key mode to use for the application.
            type: string
            enum:
                - SHARED
                - EXCLUSIVE
                - UNSPECIFIED
...
        MetadataMap:
            description: Map of Map of Object
            type: object
            additionalProperties:
                type: object
                additionalProperties:
                    type: object
...
        PageRevisionId:
            properties:
                pageId:
                    type: string
                revision:
                    type: integer
                    format: int32
...
        Links:
            properties:
                self:
                    type: string
                first:
                    type: string
                last:
                    type: string
                prev:
                    type: string
                next:
                    type: string
...
        ApplicationLinks:
            properties:
                self:
                    type: string
                members:
                    type: string
                notifications:
                    type: string
                picture:
                    type: string
                background:
                    type: string
...
        SubscriptionConsumerConfiguration:
            type: object
            description: Consumer configuration associated to the subscription in case it is attached to a push plan.
            properties:
                entrypointId:
                    type: string
                    description: The id of the targeted entrypoint
                    example: webhook
                channel:
                    type: string
                    description: The channel to consume
                    example: /channel1
                entrypointConfiguration:
                    type: object
                    description: The configuration to use at subscription time to push to the target service.
                    example: |
                        {
                            "callbackUrl": "https://webhook.site/aaaaaaaa-ffff-ffff-ffff-aaaaaaaa",
                            "headers": [
                                {
                                    "name": "demoHeader",
                                    "value": "my-value"
                                },
                                {
                                    "name": "anotherHeader",
                                    "value": "my-value2"
                                }
                            ]
                        }
            required: [entrypointId]
...
    responses:
        InternalServerError:
            description: Internal Server Error
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        PermissionError:
            description: Insufficient rights to access this resource
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
...
    securitySchemes:
        BasicAuth:
            type: http
            scheme: basic
        CookieAuth:
            type: apiKey
            in: cookie
            name: Auth-Graviteeio-APIM